fixed clicking post bug, improved page_pilot and make_post
page_pilot: clicking 4th post directly, rather than first post and skipping to the 4th. Fixed bug by centering the element on the page so it does not error any more on some pages.

Added exception to make_post; checking if file is on disk, if not readding the post to the csv with empty download_path and downloaded reset to 0 so that it will redownload on the next download run.